:star2: feat(app): add add-blip button to both views, move settings to nav
#119
