+++ app/src/backend/firebase/store/user.ts
@@ -39,7 +39,7 @@ const actions = (appConfig): ActionTree<UserState, RootState> => ({ // eslint-di
   },
   async logout ({ commit }): Promise<any> {
     commit('blips/setLoading', true, { root: true })
-    commit('user/loginState', LoginState.LOGOUT_PENDING, { root: true })
+    commit('user/setLoginState', LoginState.LOGOUT_PENDING, { root: true })
     await firebase.auth().signOut()
     commit('blips/reset', null, { root: true })
     commit('user/reset', null, { root: true })
