+++ app/src/components/app/LoadingIndicator.vue
